interface uint fixes

+++ contracts/interfaces/ITierManager.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.0;
 
 interface ITierManager {
-    function checkTierUpgrade(uint8[5] memory tierCounts) external returns (bool);
+    function checkTierUpgrade(uint32[5] memory tierCounts) external returns (bool);
     function getTokenURI(uint32 tokenId) external view returns (string memory tokenURI);
 }
